+++ regression_model/train_pipeline.py
@@ -1,27 +1,9 @@
 import numpy as np
-import pandas as pd
 from sklearn.model_selection import train_test_split
-import joblib
 
 import pipeline
-# from .data_management import load_dataset, save_pipeline
+from data_management import load_dataset, save_pipeline
 import config
-
-
-def load_dataset(*, file_name: str
-                 ) -> pd.DataFrame:
-    _data = pd.read_csv(f'{config.DATASET_DIR}/{file_name}')
-    return _data
-
-
-def save_pipeline(*, pipeline_to_persist) -> None:
-    """Persist the pipeline."""
-
-    save_file_name = 'regression_model.pkl'
-    save_path = config.TRAINED_MODEL_DIR / save_file_name
-    joblib.dump(pipeline_to_persist, save_path)
-
-    print('saved pipeline')
 
 
 def run_training() -> None:
